made restrictions with the input of calculator page, and made the navbar as component
