+++ src/index.css
@@ -146,6 +146,9 @@ button:focus-visible {
     border-radius: 25px;
 }
 
+.buttonWrapper:hover{
+    background-color: darkgray;
+}
 .calcRow {
     display: flex;
 }
@@ -194,6 +197,7 @@ button:focus-visible {
     margin-top: 20px !important;
     padding-top: 50px;
     padding-bottom: 50px;
+    width: 50%;
 }
 
 .hobbyCardHolder {
@@ -213,8 +217,8 @@ button:focus-visible {
     display: flex;
     flex-direction: column;
     flex: 1;
-    border-style: none !important;
-    height: 300px !important;
+    border-style: solid !important;
+    height: 250px !important;
     overflow: scroll !important;
 }
 
@@ -233,6 +237,11 @@ button:focus-visible {
     justify-content: space-between !important;
     flex: 1;
 }
+.hobbyBody {
+    display:flex;
+    align-items: center !important;
+    justify-content: center !important;
+}
 
 
 /* ------------------ Hobby.TSX/HobbyCard.tsx CSS ------------------*/
@@ -250,6 +259,7 @@ button:focus-visible {
     margin-top: 20px !important;
     padding-top: 50px;
     padding-bottom: 50px;
+    width: 50%;
 }
 
 .JSONCardHolder {
@@ -270,10 +280,10 @@ button:focus-visible {
     flex-direction: column;
     flex: 1;
     border-style: none !important;
-    height: 300px !important;
+    height: 200px !important;
     overflow: scroll !important;
     border-style: solid !important;
-    border-width: 1 !important;
+    text-align: left;
 }
 
 .JSONCards::-webkit-scrollbar {
@@ -285,4 +295,12 @@ button:focus-visible {
     flex-direction: column !important;
     justify-content: space-between !important;
     flex: 1;
+    justify-content: left !important;
+    align-items: left !important;
+}
+
+.JSONbody {
+    display:flex;
+    align-items: center !important;
+    justify-content: center !important;
 }
