+++ src/pages/Hobby.tsx
@@ -1,50 +1,20 @@
-import '../App.css'
-import { Link, useNavigate } from 'react-router-dom';
-import Container from 'react-bootstrap/Container';
-import Nav from 'react-bootstrap/Nav';
-import Navbar from 'react-bootstrap/Navbar';
-import NavDropdown from 'react-bootstrap/NavDropdown';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faHome } from '@fortawesome/free-solid-svg-icons';
+import '../App.css';
 import HobbyCard from '../components/HobbyCard';
 
 export default function Hobby() {
-  const navigate = useNavigate();
-
-  const handleDropdownSelect = (eventKey: string | null) => {
-    if (eventKey) {
-      navigate(eventKey);
-    }
-  };
   return (
     <>
-      <Navbar expand="lg" className="bg-body-tertiary" fixed="top" bg="dark" data-bs-theme="dark">
-          <Container>
-            <Link to="/">
-              <Navbar.Brand><FontAwesomeIcon icon={faHome} /></Navbar.Brand>
-            </Link>
-            <Navbar.Toggle aria-controls="basic-navbar-nav" />
-            <Navbar.Collapse id="basic-navbar-nav">
-              <Nav className="me-auto">
-                <NavDropdown title="" id="basic-nav-dropdown" onSelect={handleDropdownSelect}>
-                  <NavDropdown.Item eventKey="/calculator">Calculator</NavDropdown.Item>
-                  <NavDropdown.Item eventKey="/hobby">Hobbies</NavDropdown.Item>
-                  <NavDropdown.Item eventKey="/json">JSON</NavDropdown.Item>
-                </NavDropdown>
-              </Nav>
-            </Navbar.Collapse>
-          </Container>
-        </Navbar>
-  
-      <div className="hobbyWrapper">
-        <h1>SHEEEEEEEEEEEEEEEEEEESH</h1>
-        <div className="hobbyCardHolder">
-          <HobbyCard imageSrc = "./image1.jpg" hobbyName="Basketball" hobbyDesc="i love basketball"/>
-          <HobbyCard imageSrc = "./image2.jpg" hobbyName="Gaming" hobbyDesc="i love gaming"/>
-          <HobbyCard imageSrc = "./image3.jpg" hobbyName="Watching Shows/Movies" hobbyDesc="i love watchingi love watchingi love watchingi love watching"/>
-          <HobbyCard imageSrc = "./image2.jpg" hobbyName="Gaming" hobbyDesc="i love gaming"/>
-          <HobbyCard imageSrc = "./image1.jpg" hobbyName="Basketball" hobbyDesc="i love basketball"/>
-          <HobbyCard imageSrc = "./image2.jpg" hobbyName="Gaming" hobbyDesc="i love gaming"/>
+      <div className="hobbyBody">
+        <div className="hobbyWrapper">
+          <h1>SHEEEEEEEEEEEEEEEEEEESH</h1>
+          <div className="hobbyCardHolder">
+            <HobbyCard imageSrc = "./image1.jpg" hobbyName="Basketball" hobbyDesc="i love basketball"/>
+            <HobbyCard imageSrc = "./image2.jpg" hobbyName="Gaming" hobbyDesc="i love gaming"/>
+            <HobbyCard imageSrc = "./image3.jpg" hobbyName="Watching Shows/Movies" hobbyDesc="i love watchingi love watchingi love watchingi love watching"/>
+            <HobbyCard imageSrc = "./image2.jpg" hobbyName="Gaming" hobbyDesc="i love gaming"/>
+            <HobbyCard imageSrc = "./image1.jpg" hobbyName="Basketball" hobbyDesc="i love basketball"/>
+            <HobbyCard imageSrc = "./image2.jpg" hobbyName="Gaming" hobbyDesc="i love gaming"/>
+          </div>
         </div>
       </div>
     </>
