+++ src/pages/Home.tsx
@@ -1,44 +1,13 @@
-import '../App.css'
-import Form from 'react-bootstrap/Form';
-import Button from 'react-bootstrap/Button';
 import 'bootstrap/dist/css/bootstrap.min.css';
+import Button from 'react-bootstrap/Button';
 import Card from 'react-bootstrap/Card';
-import { Link, useNavigate } from 'react-router-dom';
-import Container from 'react-bootstrap/Container';
-import Nav from 'react-bootstrap/Nav';
-import Navbar from 'react-bootstrap/Navbar';
-import NavDropdown from 'react-bootstrap/NavDropdown';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faHome } from '@fortawesome/free-solid-svg-icons';
+import Form from 'react-bootstrap/Form';
+import { Link } from 'react-router-dom';
+import '../App.css';
 
 export default function Home() {
-  const navigate = useNavigate();
-
-  const handleDropdownSelect = (eventKey: string | null) => {
-    if (eventKey) {
-      navigate(eventKey);
-    }
-  };
   return (
     <>
-      <Navbar expand="lg" className="bg-body-tertiary" fixed="top" bg="dark" data-bs-theme="dark"> 
-        <Container>
-          <Link to="/">
-            <Navbar.Brand><FontAwesomeIcon icon={faHome} /></Navbar.Brand>
-          </Link>
-          <Navbar.Toggle aria-controls="basic-navbar-nav" />
-          <Navbar.Collapse id="basic-navbar-nav">
-            <Nav className="me-auto">
-              <NavDropdown title="" id="basic-nav-dropdown" onSelect={handleDropdownSelect}>
-                <NavDropdown.Item eventKey="/calculator">Calculator</NavDropdown.Item>
-                <NavDropdown.Item eventKey="/hobby">Hobbies</NavDropdown.Item>
-                <NavDropdown.Item eventKey="/json">JSON</NavDropdown.Item>
-              </NavDropdown>
-            </Nav>
-          </Navbar.Collapse>
-        </Container>
-      </Navbar>
-      
       <div className="content-wrapper">
         <h1 className="header">DIGITAL TRANSFORMATION</h1>
         <div className="nameHolder">
@@ -71,6 +40,5 @@ export default function Home() {
       
       </div>
     </>
-    
   )
 }
