+++ src/pages/JSON.tsx
@@ -1,19 +1,12 @@
-import '../App.css'
-import { Link, useNavigate } from 'react-router-dom';
-import Container from 'react-bootstrap/Container';
-import Nav from 'react-bootstrap/Nav';
-import Navbar from 'react-bootstrap/Navbar';
-import NavDropdown from 'react-bootstrap/NavDropdown';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faHome } from '@fortawesome/free-solid-svg-icons';
 import { useEffect, useState } from 'react';
+import '../App.css';
 import JSONCard from '../components/JSONCard';
 
 export default function JSON() {
 
   interface User {
     name: string;
-    userName: string;
+    username: string;
     email: string;
     address: {
       street: string;
@@ -22,15 +15,7 @@ export default function JSON() {
       zipcode: string;
     };
   }
-  
-  const navigate = useNavigate();
   const [users, setUsers] = useState<User[]>([]);
-
-  const handleDropdownSelect = (eventKey: string | null) => {
-    if (eventKey) {
-      navigate(eventKey);
-    }
-  };
 
   useEffect(() => {
     const fetchData = async () => {
@@ -49,30 +34,14 @@ export default function JSON() {
   
   return (
     <>
-    <Navbar expand="lg" className="bg-body-tertiary" fixed="top" bg="dark" data-bs-theme="dark">
-        <Container>
-          <Link to="/">
-            <Navbar.Brand><FontAwesomeIcon icon={faHome} /></Navbar.Brand>
-          </Link>
-          <Navbar.Toggle aria-controls="basic-navbar-nav" />
-          <Navbar.Collapse id="basic-navbar-nav">
-            <Nav className="me-auto">
-              <NavDropdown title="" id="basic-nav-dropdown" onSelect={handleDropdownSelect}>
-                <NavDropdown.Item eventKey="/calculator">Calculator</NavDropdown.Item>
-                <NavDropdown.Item eventKey="/hobby">Hobbies</NavDropdown.Item>
-                <NavDropdown.Item eventKey="/json">JSON</NavDropdown.Item>
-              </NavDropdown>
-            </Nav>
-          </Navbar.Collapse>
-        </Container>
-      </Navbar>
-      
-      <div className="JSONWrapper">
-        <h1>SHEEEEEEEEEEEEEEEEEEESH</h1>
-        <div className="JSONCardHolder">
-        {users.map((user) => (
-          <JSONCard key={user.name} user={user} />
-        ))}
+      <div className="JSONbody">
+        <div className="JSONWrapper">
+          <h1>SHEEEEEEEEEEEEEEEEEEESH</h1>
+          <div className="JSONCardHolder">
+          {users.map((user) => (
+            <JSONCard key={user.name} user={user} />
+          ))}
+          </div>
         </div>
       </div>
     </>
